update: fixed test_proving_worker
